Rename package to better correspond to naming conventions
