+++ .eslintignore
@@ -11,7 +11,7 @@ packages/core/helper-plugin/build/**
 packages/core/helper-plugin/lib/src/old/components/**
 packages/core/helper-plugin/lib/src/testUtils/**
 packages/core/helper-plugin/lib/src/utils/**
-packages/core/telemetry-be/lib/ampli/
+packages/core/telemetry-server/lib/ampli/
 .eslintrc.js
 .eslintrc.front.js
 .eslintrc.back.js
