+++ packages/core/content-manager/package.json
@@ -25,7 +25,7 @@
   "dependencies": {
     "@sindresorhus/slugify": "1.1.0",
     "@strapi/utils": "4.1.12",
-    "@strapi/telemetry-be": "4.1.12",
+    "@strapi/telemetry-server": "4.1.12",
     "lodash": "4.17.21"
   },
   "engines": {
