+++ packages/core/content-type-builder/package.json
@@ -31,7 +31,7 @@
     "@strapi/generators": "4.1.12",
     "@strapi/helper-plugin": "4.1.12",
     "@strapi/utils": "4.1.12",
-    "@strapi/telemetry-be": "4.1.12",
+    "@strapi/telemetry-server": "4.1.12",
     "fs-extra": "10.0.0",
     "lodash": "4.17.21",
     "pluralize": "^8.0.0",
