+++ packages/core/strapi/package.json
@@ -90,7 +90,7 @@
     "@strapi/plugin-email": "4.1.12",
     "@strapi/plugin-upload": "4.1.12",
     "@strapi/utils": "4.1.12",
-    "@strapi/telemetry-be": "4.1.12",
+    "@strapi/telemetry-server": "4.1.12",
     "bcryptjs": "2.4.3",
     "boxen": "5.1.2",
     "chalk": "4.1.2",
