+++ packages/core/telemetry-server/.editorconfig
