+++ packages/core/telemetry-server/.gitignore
