+++ packages/core/telemetry-server/.npmignore
