+++ packages/core/telemetry-server/ampli.json
