+++ packages/core/telemetry-server/lib/is-truthy.js
