+++ packages/core/telemetry-server/lib/middleware.js
