+++ packages/core/telemetry-server/lib/rate-limiter.js
