+++ packages/core/telemetry-server/lib/sender.js
