+++ packages/core/telemetry-server/lib/stringify-deep.js
