+++ packages/core/telemetry-server/lib/utils/ee.js
