+++ packages/core/telemetry-server/lib/utils/machine-id.js
