+++ packages/core/telemetry-server/lib/utils/resources/key.pub
